Move molecule-scripts to modules/
